broke down like, playlist and watchlater btns to components
